requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.13.0
Pillow==10.0.1
numpy==1.24.3
opencv-python-headless==4.8.1.78